testcase/requirements.txt
amqp==2.6.1
asgiref==3.7.2
billiard==3.6.4.0
celery==4.4.7
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
flower==0.9.7
gunicorn==21.2.0
humanize==4.7.0
kombu==4.6.11
packaging==23.1
prometheus-client==0.8.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
redis==3.5.3
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.1
vine==1.3.0
